Initial Emscripten support
Build infrastructure to compile a subset of EZ with Emscripten for web.
Currently only works with the MeshRenderSample. Still very experimental.
See "Code\BuildSystem\Web\How to build for Web.md" for details.
