+++ Code/BuildSystem/CMake/Platforms/Configure_OSX.cmake
@@ -11,7 +11,7 @@ macro(ez_platform_pull_properties)
 
 endmacro()
 
-macro (ez_platformhook_set_build_flags_clang)
+macro (ez_platformhook_set_build_flags_clang TARGET_NAME)
 
     target_compile_options(${TARGET_NAME} PRIVATE $<$<COMPILE_LANGUAGE:CXX>:-stdlib=libc++>)
 	target_link_options(${TARGET_NAME} PRIVATE $<$<COMPILE_LANGUAGE:CXX>:-stdlib=libc++>)
