+++ Code/BuildSystem/CMake/Platforms/Configure_Win.cmake
@@ -37,7 +37,7 @@ macro(ez_platform_pull_properties)
 
 endmacro()
 
-macro (ez_platformhook_set_build_flags_clang)
+macro (ez_platformhook_set_build_flags_clang TARGET_NAME)
     # Disable the warning that clang doesn't support pragma optimize.
     target_compile_options(${TARGET_NAME} PRIVATE -Wno-ignored-pragma-optimize -Wno-pragma-pack)
 endmacro()
