+++ Code/Engine/Foundation/Platform/Web/PlatformDesc_Web.cpp
@@ -1,0 +1,9 @@
+#include <Foundation/Platform/PlatformDesc.h>
+
+ezPlatformDesc g_PlatformDescWeb("Web");
+
+#if EZ_ENABLED(EZ_PLATFORM_WEB)
+
+const ezPlatformDesc* ezPlatformDesc::s_pThisPlatform = &g_PlatformDescWeb;
+
+#endif
