+++ Code/Engine/RendererFoundation/Shader/Types.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <RendererCore/RendererCoreDLL.h>
+#include <RendererFoundation/RendererFoundationDLL.h>
 
 #include <Foundation/Math/Mat3.h>
 #include <Foundation/Math/Transform.h>
@@ -11,20 +11,42 @@ class ezShaderMat3
 public:
   EZ_DECLARE_POD_TYPE();
 
+  EZ_RENDERERFOUNDATION_DLL static bool TransposeShaderMatrices /*= false*/;
+
   EZ_ALWAYS_INLINE ezShaderMat3() = default;
 
   EZ_ALWAYS_INLINE ezShaderMat3(const ezMat3& m) { *this = m; }
 
   EZ_FORCE_INLINE void operator=(const ezMat3& m)
   {
-    ezMemoryUtils::Copy(&m_Data[0], &m.m_fElementsCM[0], 3);
-    m_Data[3] = 0.0f;
+    if (ezShaderMat3::TransposeShaderMatrices)
+    {
+      m_Data[0] = m.m_fElementsCM[0];
+      m_Data[1] = m.m_fElementsCM[3];
+      m_Data[2] = m.m_fElementsCM[6];
+      m_Data[3] = 0.0f;
 
-    ezMemoryUtils::Copy(&m_Data[4], &m.m_fElementsCM[3], 3);
-    m_Data[7] = 0.0f;
+      m_Data[4] = m.m_fElementsCM[1];
+      m_Data[5] = m.m_fElementsCM[4];
+      m_Data[6] = m.m_fElementsCM[7];
+      m_Data[7] = 0.0f;
 
-    ezMemoryUtils::Copy(&m_Data[8], &m.m_fElementsCM[6], 3);
-    m_Data[11] = 0.0f;
+      m_Data[8] = m.m_fElementsCM[2];
+      m_Data[9] = m.m_fElementsCM[5];
+      m_Data[10] = m.m_fElementsCM[8];
+      m_Data[11] = 0.0f;
+    }
+    else
+    {
+      ezMemoryUtils::Copy(&m_Data[0], &m.m_fElementsCM[0], 3);
+      m_Data[3] = 0.0f;
+
+      ezMemoryUtils::Copy(&m_Data[4], &m.m_fElementsCM[3], 3);
+      m_Data[7] = 0.0f;
+
+      ezMemoryUtils::Copy(&m_Data[8], &m.m_fElementsCM[6], 3);
+      m_Data[11] = 0.0f;
+    }
   }
 
 private:
@@ -43,7 +65,20 @@ public:
 
   EZ_FORCE_INLINE void operator=(const ezMat4& m)
   {
-    ezMemoryUtils::Copy(m_Data, m.m_fElementsCM, 16);
+    if (ezShaderMat3::TransposeShaderMatrices)
+    {
+      for (ezUInt32 c = 0; c < 4; ++c)
+      {
+        m_Data[c * 4 + 0] = m.Element(0, c);
+        m_Data[c * 4 + 1] = m.Element(1, c);
+        m_Data[c * 4 + 2] = m.Element(2, c);
+        m_Data[c * 4 + 3] = m.Element(3, c);
+      }
+    }
+    else
+    {
+      ezMemoryUtils::Copy(m_Data, m.m_fElementsCM, 16);
+    }
   }
 
 private:
@@ -63,18 +98,23 @@ public:
   inline void operator=(const ezMat4& t)
   {
     float data[16];
-    t.GetAsArray(data, ezMatrixLayout::RowMajor);
 
-    for (ezUInt32 i = 0; i < 12; ++i)
-    {
-      m_Data[i] = data[i];
-    }
+    if (ezShaderMat3::TransposeShaderMatrices)
+      t.GetAsArray(data, ezMatrixLayout::ColumnMajor);
+    else
+      t.GetAsArray(data, ezMatrixLayout::RowMajor);
+
+    ezMemoryUtils::Copy(&m_Data[0], &data[0], 12);
   }
 
   inline void operator=(const ezMat3& t)
   {
     float data[9];
-    t.GetAsArray(data, ezMatrixLayout::RowMajor);
+
+    if (ezShaderMat3::TransposeShaderMatrices)
+      t.GetAsArray(data, ezMatrixLayout::ColumnMajor);
+    else
+      t.GetAsArray(data, ezMatrixLayout::RowMajor);
 
     m_Data[0] = data[0];
     m_Data[1] = data[1];
