+++ Code/Samples/MeshRenderSample/CMakeLists.txt
@@ -13,3 +13,8 @@ target_link_libraries(${PROJECT_NAME}
 )
 
 ez_add_renderers(${PROJECT_NAME})
+
+
+ez_package_files(${PROJECT_NAME} "${CMAKE_SOURCE_DIR}/Output/ShaderCache/WGSL" "Output/ShaderCache/WGSL")
+ez_package_files(${PROJECT_NAME} "${CMAKE_SOURCE_DIR}/Data/Base" "Data/Base")
+ez_package_files(${PROJECT_NAME} "${CMAKE_SOURCE_DIR}/Data/Samples/MeshRenderSample" "Data/Samples/MeshRenderSample")
