+++ Code/Samples/RtsGamePlugin/CMakeLists.txt
@@ -1,6 +1,7 @@
 ez_cmake_init()
 
 ez_requires_renderer()
+ez_requires_desktop()
 
 ez_requires_rmlui()
 ez_requires(EZ_3RDPARTY_IMGUI_SUPPORT)
