+++ Code/Samples/TextureSample/CMakeLists.txt
@@ -1,6 +1,7 @@
 ez_cmake_init()
 
 ez_requires_renderer()
+ez_requires_desktop()
 
 # Get the name of this folder as the project name
 get_filename_component(PROJECT_NAME ${CMAKE_CURRENT_SOURCE_DIR} NAME_WE)
